Tratamento de erros do CRUD

+++ src/main/java/com/SGInventario/invetario/services/ClientService.java
@@ -6,11 +6,14 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.SGInventario.invetario.dto.ClientDto;
 import com.SGInventario.invetario.entities.Client;
 import com.SGInventario.invetario.repositories.ClientRepository;
+import com.SGInventario.invetario.services.exceptions.DatabaseException;
+import com.SGInventario.invetario.services.exceptions.ResourceNotFoundException;
 
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
@@ -30,7 +33,7 @@ public class ClientService implements Serializable{
 	@Transactional
 	public ClientDto findById(Long id) {
 		Optional<Client> obj = repo.findById(id);
-		Client entity = obj.orElseThrow(() -> new RuntimeException("Id not found"));
+		Client entity = obj.orElseThrow(() -> new ResourceNotFoundException("Entity not found"));
 		return new ClientDto(entity);
 	}
 
@@ -57,18 +60,22 @@ public class ClientService implements Serializable{
 			return new ClientDto(entity);
 		}
 		catch(EntityNotFoundException e) {
-			 throw new EntityNotFoundException("Id not found: " + id);
+			 throw new ResourceNotFoundException("Id not found: " + id);
 		}
 	}
 	
 	@Transactional
 	public void delete(Long id) {
-		try {
-			repo.deleteById(id);
-		}catch(EntityNotFoundException e) {
-			throw new EntityNotFoundException("Id not found: " + id);
-		}
+	    if (!repo.existsById(id)) {
+	        throw new ResourceNotFoundException("Id not found: " + id);
+	    }
+	    try {
+	        repo.deleteById(id);
+	    } catch (DataIntegrityViolationException e) {
+	        throw new DatabaseException("Integrity violation");
+	    }
 	}
+
