+++ src/main/java/com/SGInventario/invetario/services/ProductService.java
@@ -6,11 +6,14 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.SGInventario.invetario.dto.ProductDto;
 import com.SGInventario.invetario.entities.Product;
 import com.SGInventario.invetario.repositories.ProductRepository;
+import com.SGInventario.invetario.services.exceptions.DatabaseException;
+import com.SGInventario.invetario.services.exceptions.ResourceNotFoundException;
 
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
@@ -30,7 +33,7 @@ public class ProductService implements Serializable{
 
 	public ProductDto findById(Long id) {
 		Optional<Product> obj = repo.findById(id);
-		Product entity = obj.orElseThrow(() -> new RuntimeException("Id not found"));
+		Product entity = obj.orElseThrow(() -> new ResourceNotFoundException("Entity not found"));
 		return new ProductDto(entity);
 	}
 	
@@ -60,20 +63,21 @@ public class ProductService implements Serializable{
 			repo.save(prod);
 			return new ProductDto(prod);
 		}catch (EntityNotFoundException e) {
-			throw new EntityNotFoundException("Id not found: " + id);
+			throw new ResourceNotFoundException("Id not found: " + id);
 		}
 	}
 	
 	@Transactional
-	public void delete(long id) {
-		try {
-			repo.deleteById(id);
-		}
-		catch (EntityNotFoundException e) {
-			throw new EntityNotFoundException("Id not found: " + id);
-		}
-		
+	public void delete(Long id) {
+	    if (!repo.existsById(id)) {
+	        throw new ResourceNotFoundException("Id not found: " + id);
+	    }
+	    try {
+	        repo.deleteById(id);
+	    } catch (DataIntegrityViolationException e) {
+	        throw new DatabaseException("Integrity violation");
+	    }
 	}
-	
+
 	
 }
